chore: Extract Parser to its own module (src/schema/parser.rs)
Minor internal API changes to fix the build.
No API breaks!
